requirements.txt
quadtreed3@git+https://github.com/keelezibel/quadtree-d3
wizmap
h5py
beautifulsoup4
lxml
numpy
nltk
cupy-cuda12x
spacy[cuda-autodetect,transformers,lookups]
streamlit
spacy_streamlit
tqdm
matplotlib
scikit-learn
scipy
umap-learn
pandas
ndjson
sentence_transformers